requirements.txt
# Core
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0

# Database
psycopg[binary]>=3.1,<4.0

# Async tasks
celery>=5.4,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.6,<3.0

# API docs
drf-spectacular>=0.27,<1.0

# Environment
python-decouple>=3.8,<4.0

# Production
gunicorn>=22.0,<23.0

# Monitoring
sentry-sdk[django]>=2.0,<3.0
